backend/pyproject.toml
[project]
name = "financial_advisor_api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "fastapi>=0.115.11",
    "opentelemetry-api==1.31.0",
    "opentelemetry-exporter-otlp==1.31.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.31.0",
    "opentelemetry-exporter-otlp-proto-http==1.31.0",
    "opentelemetry-instrumentation-fastapi==0.52b0",
    "opentelemetry-instrumentation-logging==0.52b0",
    "opentelemetry-instrumentation-requests==0.52b0",
    "opentelemetry-sdk==1.31.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.6.1",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "uvicorn>=0.27.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.0",
]
test = [
  "aiosqlite==0.19.0",
  "httpx==0.26.0",
  "pytest==7.4.4",
  "sqlalchemy>=2.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
asyncio_mode = "auto"

[tool.ruff]
select = ["E", "F", "I"]
line-length = 88
target-version = "py38"


